Remove DirectX stuff and do some reorganizing.
