+++ Core/Utilities/TypeConversionExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.Xna.Framework;
 
-namespace Zenith;
+namespace Zenith.Core.Utilities;
 
-public static class Utils
+public static class TypeConversionExtensions
 {
     public static Vector3 ToVector3(this Vector2 vector2) => new(vector2.X, vector2.Y, 0);
 }
