+++ ZenithMod.cs
@@ -1,4 +1,5 @@
 using Terraria.ModLoader;
+using Zenith.Core.Features.PrimitiveRendering;
 
 namespace Zenith;
 
@@ -6,6 +7,6 @@ public class ZenithMod : Mod
 {
     public override void PostSetupContent()
     {
-        ModContent.GetInstance<PrimitiveSystem>().RegisterRenderTarget("ScreenTarget");
+        ModContent.GetInstance<PrimitiveRenderingSystem>().RegisterRenderTarget("ScreenTarget");
     }
 }
